Clean up project and development dependencies
This commit cleans up the Pipfile used for this project by reducing the
dependencies declared to the minimum needed in order to run the project
through pipenv; pylint had also been marked as a development dependency
instead and bumped to version 2.6.0.

Note: For some reason, the previous Pipfile contains the colorama
dependency yet no other packages depend on it, and thus was found
appropriate to be removed.