Merge (#3)
* update (readme): change to pipenv sync

This ensures that dependencies are not updated when setting up a dev environment.

* Clean up project and development dependencies

This commit cleans up the Pipfile used for this project by reducing the
dependencies declared to the minimum needed in order to run the project
through pipenv; pylint had also been marked as a development dependency
instead and bumped to version 2.6.0.

Note: For some reason, the previous Pipfile contains the colorama
dependency yet no other packages depend on it, and thus was found
appropriate to be removed.

* Fix (reminders): fix bug on reminder error

* Feature/source command (#16)

* add source command

* feature (source): finalize source command, add brief to all commands

This feature will allow users to see the source for the github repo. I added the start to linking to exact folders for the source command. This will require more work to be complete, but this is a basic version of the source command

* update version 1.2.6

* hotfix (source command): fix no default source command

I accidentally pushed the source feature without a command for source with no arguments. This fixes that issue by returning a default embed if arg1 is None

* update (version): update version number for upcoming re-deploy

Co-authored-by: Kyler <39353605+fisher60@users.noreply.github.com>
Co-authored-by: fisher60 <wakefollower@gmail.com>
Co-authored-by: PureFunctor <purefunctor@gmail.com>