Cache doged messages
This cache ensures that the same message can't be doged multiple times.
This chec only works for messages doged since the last bot restart, would
need persistance for it to work longer than that. This bug exists due to what
I think to be an issue with Discord's emoji API.