convert memegen module to use aiohttp instead of bot session
Bot.session is deprecated and no longer works. Furthermore, async cog loading now makes it intuitive and easy to execute the async methods from the memegen module in the cog setup, rather than when the memegen class is instantiated. This also allows memegen logic to be much more modular and no longer bound to just work alongside a discord bot.