update logic for checking if user has nitro
There was a bug where we were checking if a user's pfp was animated for nitro, but a pfp could be None. I added a check to see if the pfp exists before we check if it is animated. I also added a fallback check for when no pfp exists that will pass if the user is currently boosting the server.